moved charts into separate components
